Mise en place du DELIMITER dans les messages

+++ communications/src/TTT/commons/navigation/Point.java
@@ -1,10 +1,10 @@
 package TTT.commons.navigation;
 
 public class Point {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
-	public Point(int x, int y){
+	public Point(double x, double y){
 		this.x = x;
 		this.y = y;
 	}
@@ -17,19 +17,19 @@ public class Point {
 		this(p.x,p.y);
 	}
 
-	public int getX(){
+	public double getX(){
 		return this.x;
 	}
 
-	public int getY(){
+	public double getY(){
 		return this.y;
 	}
 
-	public void setX(int x){
+	public void setX(double x){
 		this.x = x;
 	}
 
-	public void setY(int y){
+	public void setY(double y){
 		this.y = y;
 	}
